Moved modelfactory to .models namespace
