+++ sdk/communication/Azure.Communication.CallingServer/src/Models/CallingServerModelFactory.cs
@@ -3,7 +3,7 @@
 
 using Azure.Core;
 
-namespace Azure.Communication.CallingServer
+namespace Azure.Communication.CallingServer.Models
 {
     /// <summary>
     /// Calling server model factory
