+++ sdk/communication/Azure.Communication.CallingServer/src/Models/JoinCallResultInternal.cs
@@ -3,7 +3,7 @@
 
 using Azure.Core;
 
-namespace Azure.Communication.CallingServer.Models
+namespace Azure.Communication.CallingServer
 {
     [CodeGenModel("JoinCallResult")]
     internal partial class JoinCallResultInternal
